@@ -22,10 +22,9 @@ module MAdeK
 
     # Activate observers that should always be running.
     # config.active_record.observers = :cacher, :garbage_collector, :forum_observer
-    #config.active_record.identity_map = true #Rails3.1# 
-    #ActiveRecord::IdentityMap.enabled = true #Rails3.1#
+    config.active_record.identity_map = true 
 
-    config.cache_store = :mem_cache_store, {:namespace => "MAdeK_#{Rails.env}_#{Rails.root}"}
+    #Rails3.1# config.cache_store = :mem_cache_store, {:namespace => "MAdeK_#{Rails.env}_#{Rails.root}"}
 
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
@@ -36,14 +35,14 @@ module MAdeK
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
 
-    # JavaScript files you want as :defaults (application.js is always included).
-    config.action_view.javascript_expansions[:defaults] = %w(jquery.min jquery-ui.min jquery_ujs modernizr.min)
-
     # Configure the default encoding used in templates for Ruby 1.9.
     config.encoding = "utf-8"
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
+
+    # Enable the asset pipeline
+    config.assets.enabled = true
   end
 end
 
