@@ -1,4 +1,3 @@
-# -*- encoding : utf-8 -*-
 MAdeK::Application.configure do
   # Settings specified here will take precedence over those in config/environment.rb
 
@@ -12,12 +11,14 @@ MAdeK::Application.configure do
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
-  config.action_view.debug_rjs             = true #Rails3.1#
-  config.action_controller.perform_caching = true #false
+  config.action_controller.perform_caching = false 
 
   # Don't care if the mailer can't send
   config.action_mailer.raise_delivery_errors = false
 
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
+
+  # Only use best-standards-support built into browsers
+  config.action_dispatch.best_standards_support = :builtin
 end
