@@ -1,4 +1,3 @@
-# -*- encoding : utf-8 -*-
 MAdeK::Application.configure do
   # Settings specified here will take precedence over those in config/environment.rb
 
@@ -7,6 +6,10 @@ MAdeK::Application.configure do
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
+
+  # Configure static asset server for tests with Cache-Control for performance
+  config.serve_static_assets = true
+  config.static_cache_control = "public, max-age=3600"
 
   # Log error messages when you accidentally call methods on nil.
   config.whiny_nils = true
